Fix kolla_address filter plugin
